Function alcibiades::engine::run_uci

pub fn run_uci<S, T>(name: &'static str,
                     author: &'static str,
                     options: Vec<(&'static str, &'static str)>)
                     -> ! where S: DeepeningSearch<ReportData=Vec<Variation>>, T: TimeManager<S>

Runs a UCI protocol server.

"Universal Chess Interface" (UCI) is an open protocol for chess engines to communicate with other programs including Graphical User Interfaces (GUI). The protocol is independent of the operating system. For "Windows", the engine is a normal "exe" file, either a console or "real" windows application. All communication is done via standard input and output with text commands.

Parameters:

name gives the name of the engine.

author gives the name of the author.

options is a vector of (name, value) pairs that override the default configuration options.

Type parameters:

S implements game tree searching with iterative deepening. If principal variations are included in the progress reports from the search, they will be forwarded to the GUI, and eventually used to determine the best move.

Note: Normally, principal variations (PV) should be sent only when a new search depth is reached, and possibly when a new best move is found. Therefore, the majority of the progress reports will carry an empty Vec<Variation> instance. In multi-PV mode the first slot of the vector is for the best variation, the second slot is for the second-best variation, and so forth.

T is responsible for managing engine's thinking time.