Module alcibiades::stock

Implementations of various traits.

Structs

Deepening
Executes searches with iterative deepening, aspiration windows, multi-PV, and "searchmoves".

DummyTtable
Implements a transposition table that never stores anything.

SimpleEvaluator
A simple evaluator that adds a random number to the available material.

SimpleSearch
Executes depth-first alpha-beta searches with null move pruning and late move reductions.

StdMoveGenerator
Implements the MoveGenerator trait.

StdQsearch
Implements the Qsearch trait.

StdQsearchResult
Implements the QsearchResult trait.

StdSearchNode
Implements the SearchNode trait.

StdTimeManager
Implements the TimeManager trait.

StdTtable
Implements the Ttable trait.

StdTtableEntry
Implements the TtableEntry trait.