Struct alcibiades::stock::Deepening

pub struct Deepening<T: Search> { /* fields omitted */ }

Executes searches with iterative deepening, aspiration windows, multi-PV, and "searchmoves".

Iterative deepening works as follows: A depth-first search is executed with a depth of one ply, then the depth is incremented and another search is executed. This process is repeated until the search is terminated or the requested search depth is reached. In case of a terminated search, the engine can always fall back to the move selected in the last iteration of the search.

Aspiration windows are a way to reduce the search space in the search. The way it works is that we get the value from the last search iteration, calculate a window around it, and use this as alpha-beta bounds for the next search. Because the window is narrower, more beta cutoffs are achieved, and the search takes a shorter time. The drawback is that if the true score is outside this window, then a costly re-search must be made.

In multi-PV mode the engine calculates several principal variations (PV), each one starting with a different first move. This mode is very useful for chess analysis, but can make the search slower.

"searchmoves" is a feature in the UCI protocol, which makes possible to restrict the analysis to a subset of moves only. Again, this is very useful for chess analysis.

Usage

If T is a depth-first searcher, instantiate Deepening<T> to turn it into a deepening searcher with aspiration windows, multi-PV, and "searchmoves" support.

Important note: Deepening requires a proper transposition table to do its work. It can not work with DummyTtable.