Struct alcibiades::stock::StdQsearch

pub struct StdQsearch<T: MoveGenerator> { /* fields omitted */ }

Implements the Qsearch trait.

Performs classical quiescence search with stand pat, delta pruning, static exchange evaluation, check evasions, limited checks and recaptures.