Trait alcibiades::QsearchResult

pub trait QsearchResult: Clone {
    fn new(value: Value, searched_nodes: u64) -> Self;
    fn value(&self) -> Value;
    fn searched_nodes(&self) -> u64;
}

A trait for quiescence searches' results.

Required Methods

fn new(value: Value, searched_nodes: u64) -> Self

Creates a new instance.

value -- the calculated evaluation for the position. Must be between VALUE_EVAL_MIN and VALUE_EVAL_MAX.

searched_nodes -- the number of positions searched to calculate the evaluation.

fn value(&self) -> Value

Returns the calculated evaluation for the position.

Will always be between VALUE_EVAL_MIN and VALUE_EVAL_MAX.

fn searched_nodes(&self) -> u64

Retruns the number of positions searched to calculate the evaluation.