Trait alcibiades::SetOption

pub trait SetOption {
    fn options() -> Vec<(&'static str, OptionDescription)> where Self: Sized { ... }
    fn set_option(name: &str, value: &str) where Self: Sized { ... }
}

A trait for announcing and changing configuration options.

Provided Methods

fn options() -> Vec<(&'static str, OptionDescription)> where Self: Sized

Returns a list of supported configuration options (name and description).

fn set_option(name: &str, value: &str) where Self: Sized

Updates the internal state to keep up with the new value for a given configuration option.

name gives the name of the configuration option.

value is the new value for the configuration option.

Does nothing when called with unsupported configuration option name.