Module alcibiades::utils

Defines generally useful functions and types.

Structs

BoardGeometry
Tables and methods useful for move generation and position evaluation.

MoveStack
Stores a list of moves for each position in a given line of play.

ZobristArrays
Tables for calculating Zobrist hashes.

Functions

parse_fen
Parses Forsyth–Edwards Notation (FEN).

perft
Performs move path enumeration.